Added tests and refactored lazy segtree a bit
