+++ content/data-structures/LazySegmentTree.h
@@ -14,21 +14,23 @@
 
 #include "../various/BumpAllocator.h"
 
-const int inf = 1e9;
 struct Node {
+	typedef int T;
+	static constexpr T unit = -1e9;
 	Node *l = 0, *r = 0;
-	int lo, hi, mset = inf, madd = 0, val = -inf;
+	int lo, hi, mset = unit, madd = 0, val = unit;
+	T f(T a, T b) { return max(a, b); }
 	Node(int lo,int hi):lo(lo),hi(hi){} // Large interval of -inf
 	Node(vi& v, int lo, int hi) : lo(lo), hi(hi) {
 		if (lo + 1 < hi) {
 			int mid = lo + (hi - lo)/2;
 			l = new Node(v, lo, mid); r = new Node(v, mid, hi);
-			val = max(l->val, r->val);
+			val = f(l->val, r->val);
 		}
 		else val = v[lo];
 	}
 	int query(int L, int R) {
-		if (R <= lo || hi <= L) return -inf;
+		if (R <= lo || hi <= L) return unit;
 		if (L <= lo && hi <= R) return val;
 		push();
 		return max(l->query(L, R), r->query(L, R));
@@ -38,19 +40,19 @@ struct Node {
 		if (L <= lo && hi <= R) mset = val = x, madd = 0;
 		else {
 			push(), l->set(L, R, x), r->set(L, R, x);
-			val = max(l->val, r->val);
+			val = f(l->val, r->val);
 		}
 	}
 	void add(int L, int R, int x) {
 		if (R <= lo || hi <= L) return;
 		if (L <= lo && hi <= R) {
-			if (mset != inf) mset += x;
+			if (mset != unit) mset += x;
 			else madd += x;
 			val += x;
 		}
 		else {
 			push(), l->add(L, R, x), r->add(L, R, x);
-			val = max(l->val, r->val);
+			val = f(l->val, r->val);
 		}
 	}
 	void push() {
@@ -58,8 +60,8 @@ struct Node {
 			int mid = lo + (hi - lo)/2;
 			l = new Node(lo, mid); r = new Node(mid, hi);
 		}
-		if (mset != inf)
-			l->set(lo,hi,mset), r->set(lo,hi,mset), mset = inf;
+		if (mset != unit)
+			l->set(lo,hi,mset), r->set(lo,hi,mset), mset = unit;
 		else if (madd)
 			l->add(lo,hi,madd), r->add(lo,hi,madd), madd = 0;
 	}
